Added Cursor Position Retension & Time Insertion
Added functionality that will retain the cursor position when going from one line to another.
Added feature to add Current Time to the input field by pressing Insert and T.
Added comments and descriptions for functions.
Added a "saved" boolean value for each file so that it can be marked as saved or not.
Added a "symbolAutoComplete" boolean value so that the user can choose if the symbols (Curly Braces, Square Brackets, Paranthesis) should be autocompleted.
The Symbol Autocomplete functionality has not yet been implemented.