Cursor Position, Resizable Text...
When a newline is created, if the cursor position in the current line is in between the text, then the content that follows the cursor is added to the new line.
Text Size can be increased by using "Ctrl +" and "Ctrl ="
Text Size can be decreased by using "Ctrl -" and "Ctrl _"
Added a Dark Mode Indication Icon in Status Bar.
Changed App Style Colour to Green.
Changed default locale to 'en_BR' for now.