Added EOL Display, Improved Line changer
The active line changing function now have more control on the number of lines skipped. It can be configured to increase or decrease the desired number of lines.
Added an indicator on the END OF LINE format. That is, show if the End of Line is in Windows CRLF (You people are crazy by the way. 2 Characters for a newline? Let's move on!) format or Unix LF format.
Added function to change End of Line, but it is not implemented yet. And a keybinding has yet to be assigned to this action.
Added EOL value in file object. Added an application-wide EOL value too.
Order of items in the Status bar doesn't look right at all. Will have to work on that.

Quite a lot of work in 1 Commit.