Changed project name from Typecaster to Codecraft
Changed project name as I was never satisfied with the name "typecaster".
And without context, it might sound offensive. The name "typecaster" is supposed to represent Type Conversion, as in converting from one datatype to another. It is a known term in programming.
The idea was that this application is aimed to provide a platform where people's idea is converted into design, into a prototype. The idea materializes.
But it is too symbolic to be effective, and the name "codecraft" have been in my mind for so long.