Added File Save feature
The ability to save the active file have been added to the application.
It will also check if the End of Line format of the file is LF, CRLF or system and save the file accordingly.
The app will assign the directory path to the path value in the default startup file of the application.