Added Autosave, fixed Timers
Added Autosave feature for files and the application will create a directory ".codecraft" for autosave files if there does  not exist one.
Fixed the code for Timers, as it was unnecessarily individually updated for change, change in second and change in minute. Only update on change was necessary. This i not actually a fix, more like code cleanup.
Added default character encoding for default startup file as UTF-8.