Added partial Tab Character support
Added ability to enter tab character as multiple spaces into the TextField, but it is not permanent, as it only appears in the TextField and does not update the contents of the file.
But that is by design, as it will help when the display of the file contents in edit mode is modified to support code highlighting and derivation of syntax trees.
Changed the shortcut that changes active line to 'Ctrl J' and 'Ctrl M' for better access and typing speed.