Changed the Tab key handler
Pressing tab key now enters '\t' character to the textfield and updates the file content.
The Text widget doesn't display tab character properly, so I will have to implement custom individual character display in edit mode.
The Tab key also shifts the focus from the textfield which is a major problem. Pressing it again brings the focus back, but that is not a solution.