Added Basic Autocompletion with selection support
Added a basic version of autocompletion that supports selection in the TextField.
If there is an active selection in the TextField, the ' " ( [ { < characters are entered to the start of the selection,
and the ' " ) ] } > characters are added to the end of the selection.
If there is no selection, the < and > characters will not be autocompleted.