Added a basic File Explorer
A basic file explorer has been added to the application.
Although, at this point, it is not much of an explorer as it is a directory content viewer.
Currently it just displays the content of the current working directory.
The navigation to the file explorer is done by the shortcut "Ctrl F".