Added file opening capabilities
The application can now open files through the cli, using custom commands.
The filename has to be passed as the first argument while executing the application through the command line.
It will open the file, and will set the parent directory of the file as the current working directory of the application.
Also made some improvements to the file name and extension parsing logic.