Added History feature to File Explorer
The file explorer now has a History which gets updated as the user navigates through the folders.
If the user is already in the history stack navigating, any new navigation will overwrite all the history elements from the next index to the last index in the history.
Kind of like how undo works.